+++ streamer/controller_node.py
@@ -115,8 +115,8 @@ class ControllerNode(object):
       # above.
       _check_version('ffprobe', ['ffprobe', '-version'], (4, 1))
 
-      # Check that Shaka Packager version is 2.1 or above.
-      _check_version('Shaka Packager', ['packager', '-version'], (2, 1))
+      # Check that Shaka Packager version is 2.4 or above.
+      _check_version('Shaka Packager', ['packager', '-version'], (2, 4))
 
       if bucket_url:
         # Check that the Google Cloud SDK is at least v212, which introduced