+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add `--skip_deps_check` to bypass version checks on dependencies
  - Increase preserved segments outside of the availability window, improving HLS
    playback in Shaka Player
+ - Require Shaka Packager v2.4+
 
 
 ## 0.3.0 (2019-10-18)